requirements.txt
# LinguaTravel - Language Learning Assistant Requirements
# Python 3.8+ required

# Core UI Framework
gradio>=4.0.0

# Speech Recognition
openai-whisper>=20231117

# HTTP Client for Ollama API
requests>=2.31.0

# Environment Configuration
python-dotenv>=1.0.0

# Audio Processing (required by Whisper)
ffmpeg-python>=0.2.0

# Deep Learning Framework (required by Whisper)
torch>=2.0.0
torchaudio>=2.0.0

# Numerical Computing
numpy>=1.24.0

# Additional Whisper Dependencies
tiktoken>=0.5.0
more-itertools>=10.0.0
